requirements.txt
Flask==3.0.3
Flask_Cors==4.0.0
langchain_community==0.2.4
langchain_core==0.2.6
langchain_openai==0.1.8
langchain_postgres==0.0.7
pandas==2.2.2
Pillow==10.3.0
psycopg2_binary==2.9.9
python-dotenv==1.0.1
